+++ routes/index.js
@@ -1,7 +1,6 @@
 const express = require('express')
 const router = express.Router()
 const config = require('../config')
-// const Dog = require('../models/dog');
 
 router.use((req, res, next) => {
   res.locals = config
@@ -9,23 +8,23 @@ router.use((req, res, next) => {
 })
 
 router.get('/', (req, res) => {
-  res.render('../views/pages/index', {pageTitle: "Index"})
+  res.render('pages/index', {pageTitle: "Index"})
 })
 
 router.get('/subscribe', (req, res) => {
-  res.render('../views/pages/subscribe', {pageTitle: "Subscribe"})
+  res.render('pages/subscribe', {pageTitle: "Subscribe"})
 })
 
 router.get('/team', (req, res) => {
-  res.render('../views/pages/team', {pageTitle: "Team Members"})
+  res.render('pages/team', {pageTitle: "Team Members"})
 })
 
 router.get('/admin', (req, res) => {
-  res.render('../views/pages/admin', {pageTitle: "Subscribers List"})
+  res.render('pages/admin', {pageTitle: "Subscribers List"})
 })
 
 router.get('/gallery', (req, res) => {
-  res.render('../views/pages/gallery', {pageTitle: "Gallery"})
+  res.render('pages/gallery', {pageTitle: "Gallery"})
 })
 
 module.exports = router;
