+++ server.js
@@ -3,18 +3,23 @@ const mongoose = require('./_connection')
 const dotenv = require('dotenv').config()
 const express = require('express')
 const apiS = require('./routes/api/apiSubscribers')
-// const subscribe = require('./routes/subscribe')
 const index = require('./routes/index')
+const sub = require('./routes/subscribed')
+// const team = require('./routes/team')
+// const admin = require('./routes/admin')
+// const gallery = require('./routes/gallery')
+
 
 const app = express()
 
 app.set('view engine', 'ejs');
-
 app.use(express.static('./public'))
 
-app.use('/', index)
 app.use('/api/v0', apiS)
-// app.use('/subscribe', subscribe)
+app.use('/subscribed', sub)
+// app.use('/team', team)
+
+app.use('/', index)
 
 app.use((req, res) => {
   res.sendStatus(404);
