+++ views/pages/subscribe.ejs
@@ -12,7 +12,7 @@
 
     </head>
     <main>
-      <form action="/subscribers" method="post">
+      <form action="/subscribed" method="post">
         <fieldset>
           <legend>Follow us to get newest updates:</legend>
           <p>
