+++ views/pages/subscribed.ejs
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en-US">
+  <head>
+    <title>Subscribed</title>
+		<meta charset="utf-8" />
+		<meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no" />
+		<link rel="stylesheet" href="css/main.css" />
+    <link rel="stylesheet" href="css/subscribe.css">
+  </head>
+  <body>
+    <head>
+
+    </head>
+    <main>
+      <h1>yeah!</h1>
+    </main>
+    <footer>
+
+    </footer>
+  </body>
+</html>
